refactor: 패키지 구조 변경
- admin, home, common 패키지 생성
- header 멘트 변경
