+++ front/src/App.js
@@ -1,9 +1,9 @@
 import React from "react";
 import { Route, Switch } from "react-router-dom";
 import "./index.css";
-import Header from "./component/Header";
-import Main from "./component/Main";
-import Home from "./component/Home";
+import Header from "./component/common/Header";
+import Main from "./component/admin/Main";
+import Home from "./component/home/Home";
 
 const App = () => {
   return (
