+++ front/src/component/common/Header.js
@@ -1,12 +1,12 @@
 import React from "react";
-import "../css/header.css";
+import "../../css/common/header.css";
 
 const Header = () => {
   return (
     <div className="headerContainer">
-      <div className="header">칵테일픽 관리자 페이지 - 칵테일</div>
+      <div className="header">좋은 술을 고민 없이, 칵테일픽</div>
     </div>
   );
-}
+};
 
 export default Header;
